+++ accounting/RELEASES.md
@@ -79,3 +79,13 @@ Based on:
 - [python v0.14.1] accounting
 ### Releases
 - [PyPI v0.14.1] https://pypi.org/project/apideck-accounting-unify/0.14.1 - accounting
+
+## 2025-05-26 00:05:20
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.551.0 (2.610.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.15.0] accounting
+### Releases
+- [PyPI v0.15.0] https://pypi.org/project/apideck-accounting-unify/0.15.0 - accounting